Add Reset to File menu when booted from floppy.  Add RadioBox to choose level of mtutor to run on-line.  Have mtutor give up thread for 1 msec on call to r.input if booted from floppy.
git-svn-id: svn://akdesign.dyndns.org/dtcyber/branches/floppy@2166 d8f3d347-7192-4e35-a67b-6d701703d080

+++ Z80.h
@@ -20,7 +20,7 @@ public:
 
     bool m_giveup8080;
     bool m_mtutorPatch;
-    unsigned short  m_mtPLevel;
+    long m_mtPLevel;
     bool m_mtutorBoot;
     bool m_MtTrace;
     unsigned short  m_zclock;      // one second clock for mtutor
